+++ pkg/clients/docker_tasks_container_create_test.go
@@ -87,6 +87,10 @@ func setupContainerMocks() (*clients.MockDocker, *clients.ImageLog) {
 	md.On("NetworkConnect", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(nil)
 	md.On("NetworkDisconnect", mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(nil)
 
+	md.On("VolumeList", mock.Anything, mock.Anything).Return(nil, nil)
+	md.On("VolumeCreate", mock.Anything, mock.Anything).Return(types.Volume{Name: "test_volume"}, nil)
+	md.On("VolumeRemove", mock.Anything, mock.Anything, mock.Anything).Return(nil)
+
 	mic := &clients.ImageLog{}
 	mic.On("Log", mock.Anything, mock.Anything).Return(nil)