+++ pkg/clients/mocks/container_tasks.go
@@ -71,10 +71,14 @@ func (d *MockContainerTasks) CopyFromContainer(id, src, dst string) error {
 	return args.Error(0)
 }
 
-func (d *MockContainerTasks) CopyLocalDockerImageToVolume(images []string, volume string) (string, error) {
+func (d *MockContainerTasks) CopyLocalDockerImageToVolume(images []string, volume string) ([]string, error) {
 	args := d.Called(images, volume)
 
-	return args.String(0), args.Error(1)
+	if a, ok := args.Get(0).([]string); ok {
+		return a, args.Error(1)
+	}
+
+	return nil, args.Error(1)
 }
 
 func (d *MockContainerTasks) ExecuteCommand(id string, command []string, writer io.Writer) error {