+++ pkg/providers/cluster_k3s.go
@@ -60,7 +60,7 @@ func (c *K8sCluster) Destroy() error {
 
 // Lookup the a clusters current state
 func (c *K8sCluster) Lookup() ([]string, error) {
-	return c.client.FindContainerIDs(c.config.Name, c.config.Type)
+	return c.client.FindContainerIDs(fmt.Sprintf("server.%s", c.config.Name), c.config.Type)
 }
 
 func (c *K8sCluster) createK3s() error {
@@ -86,7 +86,7 @@ func (c *K8sCluster) createK3s() error {
 	}
 
 	// create the volume for the cluster
-	volID, err := c.client.CreateVolume(c.config.Name)
+	volID, err := c.client.CreateVolume("images")
 	if err != nil {
 		return err
 	}
@@ -183,7 +183,7 @@ func (c *K8sCluster) createK3s() error {
 	// import the images to the servers container d instance
 	// importing images means that k3s does not need to pull from a remote docker hub
 	if c.config.Images != nil && len(c.config.Images) > 0 {
-		err := c.ImportLocalDockerImages(c.config.Name, id, c.config.Images)
+		err := c.ImportLocalDockerImages("images", id, c.config.Images)
 		if err != nil {
 			return xerrors.Errorf("Error importing Docker images: %w", err)
 		}
@@ -287,16 +287,18 @@ func (c *K8sCluster) ImportLocalDockerImages(name string, id string, images []co
 
 	// import to volume
 	vn := utils.FQDNVolumeName(name)
-	imageFile, err := c.client.CopyLocalDockerImageToVolume(imgs, vn)
+	imagesFile, err := c.client.CopyLocalDockerImageToVolume(imgs, vn)
 	if err != nil {
 		return err
 	}
 
-	// execute the command to import the image
-	// write any command output to the logger
-	err = c.client.ExecuteCommand(id, []string{"ctr", "image", "import", "/images/" + imageFile}, c.log.StandardWriter(&hclog.StandardLoggerOptions{ForceLevel: hclog.Debug}))
-	if err != nil {
-		return err
+	for _, i := range imagesFile {
+		// execute the command to import the image
+		// write any command output to the logger
+		err = c.client.ExecuteCommand(id, []string{"ctr", "image", "import", "/images/" + i}, c.log.StandardWriter(&hclog.StandardLoggerOptions{ForceLevel: hclog.Debug}))
+		if err != nil {
+			return err
+		}
 	}
 
 	return nil
@@ -325,6 +327,8 @@ func (c *K8sCluster) destroyK3s() error {
 		}
 	}
 
+	return nil
+
 	// delete the volume
-	return c.client.RemoveVolume(c.config.Name)
+	//return c.client.RemoveVolume(c.config.Name)
 }